renamed seqfold to fold.
